iocuddle
Struct Ioctl

#[repr(transparent)]
pub struct Ioctl<D, T>(pub(crate) c_ulong, pub(crate) PhantomData<(D, T)>);

A defined ioctl along with its associated direction and type

Tuple Fields
0: c_ulong1: PhantomData<(D, T)>
Implementations

impl<D, T> Ioctl<D, T>

pub const unsafe fn classic(request: c_ulong) -> Self

Define a new ioctl with an associated direction and type

The request argument is the allocated integer which uniquely identifies this ioctl.

Safety

This function is unsafe because defining an ioctl with an incorrect request, direction or argument type can later result in memory corruption. You are responsible to ensure these values are correct.

Further, you are responsible to ensure that the argument type itself provides appropriate safe wrappers around its raw contents. For some types none are required. For others, particularly types that pass pointers to the kernel as u64, you need to ensure that things like lifetimes are correct.

pub const unsafe fn lie<E, U>(self) -> Ioctl<E, U>

Lie about the ioctl direction or type

This function should be avoided unless necessary.

Sometimes kernel developers make mistakes and use the wrong macros or types during their ioctl definitions. However, once merged these form part of the userspace API and wonโt be broken. Therefore, we need a way to use the request number with the incorrect type. This function allows this.

Whenever using this function, it would be wise to include a comment with a link to the kernelโs ioctl definition and explaining why the definition is incorrect.

Safety

For safety details, see Ioctl::classic.

Additionally, one should note that this function discards your normal protections. So you need to make sure that you have it correct.

impl Ioctl<Read, c_void>

pub fn ioctl(self, fd: &impl AsRawFd) -> Result<c_uint>

Issue an ioctl to read a file descriptorโs metadata as c_uint.

No argument is supplied to the internal ioctl() call. The raw (positive) return value from the internal ioctl() call is returned on success.

impl<T> Ioctl<Read, &T>

pub fn ioctl(self, fd: &impl AsRawFd) -> Result<(c_uint, T)>

Issue an ioctl to read a file descriptorโs metadata as type T.

A zeroed instance of type T is passed as the first argument to the internal ioctl() call. Upon success, returns the raw (positive) return value and the instance of T.

impl Ioctl<Write, c_void>

pub fn ioctl(self, fd: &mut impl AsRawFd) -> Result<c_uint>

Issue an ioctl to modify a file descriptor

No argument is provided.

On success, returns the (positive) return value.

impl Ioctl<Write, c_int>

pub fn ioctl(self, fd: &mut impl AsRawFd, data: c_int) -> Result<c_uint>

Issue an ioctl to modify a file descriptor

A C-integer argument is provided.

On success, returns the (positive) return value.

impl<T> Ioctl<Write, &T>

pub fn ioctl(self, fd: &mut impl AsRawFd, data: &T) -> Result<c_uint>

Issue an ioctl to modify a file descriptor

A reference to an immutable instance of T is provided as the argument.

On success, returns the (positive) return value.

impl<T> Ioctl<WriteRead, &T>

pub fn ioctl(self, fd: &mut impl AsRawFd, data: &mut T) -> Result<c_uint>

Issue an ioctl to modify a file descriptor and read its metadata

A reference to a mutable instance of T is provided as the argument.

On success, returns the (positive) return value.